rename types.hpp -> stf.hpp
