+++ code/tests/stf/cpp/stf/alg/intersection_tests.cpp
@@ -2,7 +2,7 @@
 
 #include <gtest/gtest.h>
 
-#include <stf/types.hpp>
+#include <stf/stf.hpp>
 
 #include "stf/alg/scaffolding/intersection.hpp"
