+++ code/tests/stf/cpp/stf/cam/scamera_tests.cpp
@@ -2,7 +2,7 @@
 
 #include <gtest/gtest.h>
 
-#include <stf/types.hpp>
+#include <stf/stf.hpp>
 
 #include "stf/cam/scaffolding/scamera.hpp"
