+++ code/tests/stf/cpp/stf/geom/segment2_tests.cpp
@@ -2,7 +2,7 @@
 
 #include <gtest/gtest.h>
 
-#include <stf/types.hpp>
+#include <stf/stf.hpp>
 
 #include "stf/geom/scaffolding/segment.hpp"
