+++ code/tests/stf/cpp/stf/math/mtx4_tests.cpp
@@ -2,7 +2,7 @@
 
 #include <gtest/gtest.h>
 
-#include <stf/types.hpp>
+#include <stf/stf.hpp>
 
 namespace stf::math
 {
